refactor(military): extract the chart event rail into EventRail.svelte
Pull the DOM annotation rail (convertToPixel positioning, proximity
merging, tick rules, crest gutter, hover tooltip) out of MilitaryTab into
a reusable component so the Techs tab's science rail can share it.
MilitaryTab keeps shaping its leader/law/tech marker groups and the
on-plot highlight guide; no visual change.
